chore(linkedin): share messaging thread URL helpers
