+++ clis/linkedin/safe-send.js
@@ -1,42 +1,15 @@
 import { cli, Strategy } from '@jackwener/opencli/registry';
 import { ArgumentError, AuthRequiredError, CommandExecutionError } from '@jackwener/opencli/errors';
 import { createHash } from 'node:crypto';
-import { unwrapEvaluateResult } from './shared.js';
+import { canonicalizeLinkedInThreadUrl, normalizeWhitespace, unwrapEvaluateResult } from './shared.js';
 
 const LINKEDIN_DOMAIN = 'www.linkedin.com';
-
-function normalizeWhitespace(value) {
-  return String(value ?? '').replace(/[\u00a0\u202f]/g, ' ').replace(/\s+/g, ' ').trim();
-}
 
 function normalizeName(value) {
   return normalizeWhitespace(value)
     .replace(/\s*[•·]\s*(?:1st|2nd|3rd\+?|degree connection).*$/i, '')
     .replace(/\s+LinkedIn.*$/i, '')
     .toLowerCase();
-}
-
-function isLinkedInHost(hostname) {
-  const host = String(hostname || '').toLowerCase();
-  return host === 'linkedin.com' || host.endsWith('.linkedin.com');
-}
-
-function canonicalizeLinkedInThreadUrl(value) {
-  const raw = normalizeWhitespace(value);
-  if (!raw) return '';
-  try {
-    const url = new URL(raw);
-    if (url.protocol !== 'https:' || url.username || url.password || url.port || !isLinkedInHost(url.hostname)) return '';
-    const match = url.pathname.match(/^\/messaging\/thread\/([^/]+)\/?$/i);
-    if (!match || !match[1]) return '';
-    url.hostname = 'www.linkedin.com';
-    url.hash = '';
-    url.search = '';
-    if (!url.pathname.endsWith('/')) url.pathname += '/';
-    return url.toString();
-  } catch {
-    return '';
-  }
 }
 
 function hashText(value) {
@@ -344,9 +317,7 @@ cli({
 });
 
 export const __test__ = {
-  normalizeWhitespace,
   normalizeName,
-  canonicalizeLinkedInThreadUrl,
   hashText,
   assessThreadSafety,
 };
